:wrench: Create a common editor class
