+++ pyflow/core/pyeditor.py
@@ -14,6 +14,7 @@ from PyQt5.QtGui import (
     QWheelEvent,
 )
 from PyQt5.Qsci import QsciScintilla, QsciLexerPython
+from pyflow.core.editor import Editor
 from pyflow.graphics.theme_manager import theme_manager
 
 from pyflow.blocks.block import Block
@@ -24,7 +25,7 @@ if TYPE_CHECKING:
 POINT_SIZE = 11
 
 
-class PythonEditor(QsciScintilla):
+class PythonEditor(Editor):
 
     """In-block python editor for Pyflow."""
 
@@ -35,9 +36,7 @@ class PythonEditor(QsciScintilla):
             block: Block in which to add the python editor widget.
 
         """
-        super().__init__(None)
-        self._mode = "NOOP"
-        self.block = block
+        super().__init__(block)
 
         self.update_theme()
         theme_manager().themeChanged.connect(self.update_theme)
@@ -89,35 +88,13 @@ class PythonEditor(QsciScintilla):
         lexer.setFont(font)
         self.setLexer(lexer)
 
-    def views(self) -> List["View"]:
-        """Get the views in which the python_editor is present."""
-        return self.block.scene().views()
-
     def wheelEvent(self, event: QWheelEvent) -> None:
         """How PythonEditor handles wheel events."""
         if self.mode == "EDITING" and event.angleDelta().x() == 0:
             event.accept()
             return super().wheelEvent(event)
 
-    @property
-    def mode(self) -> int:
-        """PythonEditor current mode."""
-        return self._mode
-
-    @mode.setter
-    def mode(self, value: str):
-        self._mode = value
-        for view in self.views():
-            view.set_mode(value)
-
-    def mousePressEvent(self, event: QMouseEvent) -> None:
-        """PythonEditor reaction to PyQt mousePressEvent events."""
-        if event.buttons() & Qt.MouseButton.LeftButton:
-            self.mode = "EDITING"
-        return super().mousePressEvent(event)
-
     def focusOutEvent(self, event: QFocusEvent):
         """PythonEditor reaction to PyQt focusOut events."""
-        self.mode = "NOOP"
         self.block.source = self.text()
         return super().focusOutEvent(event)
